% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{executions}
\alias{executions}
\title{USA Executions Since Gregg v. Georgia (1976)}
\format{
\subsection{\code{executions}}{

A data frame with 2332 rows and 12 columns:
\describe{
\item{victim_id}{Unique ID of a victim of an individual who was executed}
\item{victim_race}{The race of the victim}
\item{victim_sex}{The sex of the victim}
\item{execution_id}{Unique ID of the individual who was executed, the perpetrator}
\item{perpetrator_name}{Name of the perpetrator}
\item{perpetrator_race}{Race of the perpetrator}
\item{perpetrator_sex}{Sex of the perpetrator}
\item{perpetrator_age}{Age of the perpetrator}
\item{state_of_execution}{State in which the execution occurred}
\item{federal_execution_flag}{Whether or not the execution was federal}
\item{execution_date}{Date that the execution occurred}
\item{execution_method}{Method used for the execution}
}
}
}
\source{
\url{https://deathpenaltyinfo.org/facts-and-research/data/executions}
}
\usage{
executions
}
\description{
This dataset contains a list of executions in the United States since the
1976 Supreme Court decision in Gregg v. Georgia, which reinstated the death penalty
in the U.S. The dataset includes details about the victims, perpetrators, and
execution methods used. It is a cleaned and abbreviated version of the original
dataset, which can be obtained at the source link below.
}
\keyword{datasets}
