% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ny_shooting}
\alias{ny_shooting}
\title{Shooting Incidents in New York City (2006-2024)}
\format{
\subsection{\code{ny_shooting}}{

A data frame with 23298 rows and 17 columns:
\describe{
\item{incident_id}{Unique ID of the shooting incident}
\item{occur_date}{Date of the shooting incident}
\item{occur_time}{Time of the shooting incident}
\item{borough}{Borough in which the shooting incident occurred}
\item{precinct}{Identifier of the precinct of the shooting incident}
\item{location_class}{Classification of the location where the shooting incident occurred}
\item{location_description}{Description of the location where the shooting incident occurred}
\item{indoor_flag}{Indicator of whether the shooting occurred indoors}
\item{latitude}{Latitude of the location of the shooting incident}
\item{longitude}{Longitude of the location of the shooting incient}
\item{victim_age_group}{Age group of the victim involved the shooting incident}
\item{victim_sex}{Sex of the victim involved in the shooting incident}
\item{victim_race_ethnicity}{Race/ethnicity of the victim involved in the shooting incident}
\item{perpetrator_age_group}{Age group of the perpetrator}
\item{perpetrator_sex}{Sex of the perpetrator}
\item{perpetrator_race_eth}{Race/ethnicity of the perpetrator}
\item{murder_flag}{Indicator of whether the shooting incident was a murder}
}
}
}
\source{
\url{https://catalog.data.gov/dataset/nypd-shooting-incident-data-historic}
}
\usage{
ny_shooting
}
\description{
This dataset provides a breakdown of every shooting incident that occurred
in NYC from 2006 through the end of the previous calendar year. The data
is manually extracted every quarter and reviewed by the Office of Management
Analysis and Planning before being posted on the NYPD website. Each record
represents a shooting incident in NYC and includes information about the
event, the location, and the time of occurrence. Additionally, information
related to suspect and victim demographics is also included. This dataset
is a cleaned and abbreviated version of the original NYPD shooting incident
data, which can be obtained at the source link below.
}
\keyword{datasets}
