% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{students}
\alias{students}
\title{NCES High School Study 2009}
\format{
\subsection{\code{students}}{

A data frame with 23503 rows and 37 columns:
\describe{
\item{student_id}{Unique student identifier}
\item{survey1_date}{Month and year of the base survey}
...
}
}
}
\source{
\url{https://nces.ed.gov/surveys/hsls09}
}
\usage{
students
}
\description{
High school longitudinal study (add more later)
}
\keyword{datasets}
